frontend/requirements.txt
flask
kafka-python
gunicorn
